UPDATE schemas/carbon_footprint.py - the spec actually defines this field as decimal, fixed pagination

+++ schemas/carbon_footprint.py
@@ -7,8 +7,8 @@ from pydantic import field_validator, BaseModel, Field
 class CarbonFootprint(BaseModel):
     declaredUnit: str
     unitaryProductAmount: int
-    pCfExcludingBiogenic: int
-    pCfIncludingBiogenic: int
+    pCfExcludingBiogenic: float
+    pCfIncludingBiogenic: float
     fossilGhgEmissions: int
     fossilCarbonContent: int
     biogenicCarbonContent: int
